+++ csharp-ms/SupermarketReceipt/OfferTypes/OfferHandlerBase.cs
@@ -12,6 +12,6 @@ namespace SupermarketReceipt.OfferTypes
         {
             this.next = next;
         }
-        public abstract void HandleOffer(Receipt receipt, Dictionary<Product, Offer> specialOffers, Product product, int quantity, SupermarketCatalog catalog);
+        public abstract void HandleOffer(Receipt receipt, Dictionary<Product, Offer> specialOffers, Product product, double quantity, SupermarketCatalog catalog);
     }
 }
